# Move VM global tables from C to ML