--- a/kernel/kernel.mllib
+++ b/kernel/kernel.mllib
@@ -43,6 +43,6 @@ Subtyping
 Mod_typing
 Nativelibrary
 Safe_typing
-Vm
 Csymtable
+Vm
 Vconv