--- a/kernel/vmvalues.ml
+++ b/kernel/vmvalues.ml
@@ -43,6 +43,7 @@ let fix_val v = (Obj.magic v : values)
 let cofix_upd_val v = (Obj.magic v : values)
 
 type vm_env
+type vm_global
 let fun_env v = (Obj.magic v : vm_env)
 let fix_env v = (Obj.magic v : vm_env)
 let cofix_env v = (Obj.magic v : vm_env)
@@ -50,6 +51,8 @@ let cofix_upd_env v = (Obj.magic v : vm_env)
 type vstack = values array
 
 let fun_of_val v = (Obj.magic v : vfun)
+
+let vm_global (v : values array) = (Obj.magic v : vm_global)
 
 (*******************************************)
 (* Machine code *** ************************)