--- a/kernel/vmvalues.mli
+++ b/kernel/vmvalues.mli
@@ -15,6 +15,7 @@ open Cbytecodes
 
 type values
 type vm_env
+type vm_global
 type vprod
 type vfun
 type vfix
@@ -32,6 +33,8 @@ val fun_env : vfun -> vm_env
 val fix_env : vfix -> vm_env
 val cofix_env : vcofix -> vm_env
 val cofix_upd_env : to_update -> vm_env
+
+val vm_global : values array -> vm_global
 
 (** Cast a value known to be a function, unsafe in general *)
 val fun_of_val : values -> vfun