This PR moves the VM global tables from the C implementation to the ML implementation. We declare them in ML code and pass them as an argument to the C reduction functions. This is a first step towards being able to synchronize this state with the Coq summary, so as to free unused memory, as well as locally replacing the value of a global, e.g. to implement local opacity in the VM.

This also makes the code slightly cleaner, e.g. automatically properly initializing the rel tables without them containing dummy data.

I was afraid that this might have been slower due to the C compiler performing less optimizations, but it doesn't seem to be the case after all. I guess it makes sense, as the pointer is allocated by the OCaml GC anyways.

Bench:
```
┌──────────────────────────┬─────────────────────────┬─────────────────────────────────────┬───────────────────────────────────────┬─────────────────────────┬─────────────────┐
│                          │      user time [s]      │             CPU cycles              │           CPU instructions            │  max resident mem [KB]  │   mem faults    │
│                          │                         │                                     │                                       │                         │                 │
│             package_name │     NEW     OLD PDIFF   │           NEW           OLD PDIFF   │            NEW            OLD PDIFF   │     NEW     OLD PDIFF   │ NEW OLD PDIFF   │
├──────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│    coq-mathcomp-fingroup │   60.44   60.80 -0.59 % │  166428584127  166896752508 -0.28 % │   231469653541   231593430204 -0.05 % │  529468  529564 -0.02 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│   coq-mathcomp-odd_order │ 1338.04 1343.05 -0.37 % │ 3722799312064 3730030746943 -0.19 % │  6591138192844  6588931938305 +0.03 % │ 1002808 1074176 -6.64 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│   coq-mathcomp-character │  262.67  263.37 -0.27 % │  728361468299  730425121020 -0.28 % │  1074466695631  1074665224701 -0.02 % │  984344  984600 -0.03 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│               coq-geocoq │ 3047.73 3051.96 -0.14 % │ 8476299150818 8486750182257 -0.12 % │ 13965502296280 13967856271799 -0.02 % │ 1240308 1242196 -0.15 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│             coq-compcert │  814.51  814.76 -0.03 % │ 2273965839896 2272410670086 +0.07 % │  3424265355466  3425603069430 -0.04 % │ 1368320 1369244 -0.07 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│     coq-mathcomp-algebra │  169.31  169.32 -0.01 % │  467079817224  466856454705 +0.05 % │   656166961721   656018828796 +0.02 % │  588608  588768 -0.03 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│         coq-fiat-parsers │  659.88  659.57 +0.05 % │ 1840217524385 1843765680640 -0.19 % │  2957382112931  2958206304060 -0.03 % │ 3374380 3348048 +0.79 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│           coq-coquelicot │   75.43   75.39 +0.05 % │  207446521549  207965394423 -0.25 % │   259630758430   259967721173 -0.13 % │  660052  656348 +0.56 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│              coq-bignums │   75.10   75.05 +0.07 % │  208687387455  208352870670 +0.16 % │   283223957401   283302640068 -0.03 % │  526356  526552 -0.04 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│                coq-color │  570.68  570.26 +0.07 % │ 1586587466967 1585912435822 +0.04 % │  1989591666968  1990298191910 -0.04 % │ 1438704 1439344 -0.04 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│                 coq-hott │  309.40  309.14 +0.08 % │  842513689729  842397799466 +0.01 % │  1375685832726  1376007763700 -0.02 % │  496448  496576 -0.03 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│ coq-mathcomp-real_closed │  165.54  165.34 +0.12 % │  459517066767  458971705800 +0.12 % │   699295227362   699416163376 -0.02 % │  714956  715232 -0.04 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│                coq-flocq │   59.36   59.23 +0.22 % │  163587367310  163808903677 -0.14 % │   211527066041   211563117381 -0.02 % │  664056  660820 +0.49 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│       coq-mathcomp-field │  441.04  439.72 +0.30 % │ 1223911680284 1220353697453 +0.29 % │  2024761406182  2024961143264 -0.01 % │  713300  713760 -0.06 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│                 coq-corn │ 1479.40 1474.02 +0.36 % │ 4106968669794 4102830785496 +0.10 % │  6397483028130  6400106610373 -0.04 % │  839512  835800 +0.44 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│    coq-mathcomp-solvable │  190.59  189.75 +0.44 % │  526621150340  526430860914 +0.04 % │   761548805325   761445630119 +0.01 % │  716420  716156 +0.04 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│            coq-fiat-core │  102.73  102.20 +0.52 % │  290549062565  290036335676 +0.18 % │   378266514772   378135129211 +0.03 % │  500864  501416 -0.11 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│   coq-mathcomp-ssreflect │   40.70   40.48 +0.54 % │  111612473841  111675287704 -0.06 % │   137656515666   137869727221 -0.15 % │  478660  478812 -0.03 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│      coq-formal-topology │   34.33   34.08 +0.73 % │   94943536842   94961475070 -0.02 % │   125133046104   125084139188 +0.04 % │  483396  483864 -0.10 % │   0   0  +nan % │
└──────────────────────────┴─────────────────────────┴─────────────────────────────────────┴───────────────────────────────────────┴─────────────────────────┴─────────────────┘


PDIFF = proportional difference between measurements done for the NEW and the OLD Coq version
      = (NEW_measurement - OLD_measurement) / OLD_measurement * 100%

NEW = 3c38a282d505ed2ced734ac2f212c9328d704b6a
OLD = df9d3a36e71d6d224286811fdc529ad5a955deb7
```