@maximedenes 2018-03-04
@ppedrot Do you plan to do other VM cleanups?