@ppedrot 2018-03-04
Not for 8.8, I think. This particular PR shouldn't really interfere with your patches, IIUC, it is quite self-contained and local.