@silene 2018-03-06
I thought the code was using a sane allocation strategy, but now that I am reading your patch, it seems like it is the worst possible one (the quadratic one). Don't you want to fix that issue along the way? Other than that, the patch is fine and welcome.