@ppedrot 2018-03-07
@silene I ran the code with an assertion failure for reallocation first, and it seems that on the stdlib, we never reallocate. I could use a O(1) amortized allocation, but actually I'm not sure this happens often in practice...

(Also, reading it carefully, there are memory leaks in the C code, and probably strict pointer aliasing breaking as well. This is making me worried about the overall safety of the VM code. If there is something I don't trust, C code is well ranked on that list.)