@silene 2018-03-07
>  there are memory leaks in the C code, and probably strict pointer aliasing breaking as well

Is that so? A quick search for `*)` did not bring anything interesting up. I would have thought that the C code always reads or writes OCaml's `value`s in and out of memory. You break pointer aliasing only when you read an object with a pointer that has a type incompatible with the pointer used to write the object.

As for memory leaks, do you mean that the compiled bytecode is not freed on backtrack? Or do you have something else in mind?