@ppedrot 2018-03-07
> As for memory leaks, do you mean that the compiled bytecode is not freed on backtrack? Or do you have something else in mind?

Bytecode is indeed never freed, but also a few globally allocated data that are reallocated on resizing of tables. Have a look at the uses of `coq_stat_alloc`.