@ppedrot 2018-03-07
> Is that so? A quick search for *) did not bring anything interesting up. I would have thought that the C code always reads or writes OCaml's values in and out of memory. You break pointer aliasing only when you read an object with a pointer that has a type incompatible with the pointer used to write the object.

I can find at least one suspicious use in `coq_interp.c`, in the `coq_interprete` function there are a few casts of the form
```
pc = (code_t)(sp[0]);
```
where `pc : code_t` and `sp : value*`. Looking at the OCaml headers, `code_t ≡ int32_t*` and `value ≡ int` or `value ≡ long` depending on the architecture.