@ppedrot 2018-03-07
Forget what I said and let's blame the jetlag, it's just a cast `value ↦ int32_t` and those ones are valid.