@silene 2018-03-07
> Have a look at the uses of `coq_stat_alloc`.

As far as I understand, `coq_makeaccu` is fine, and so is `coq_pushpop`, and so are `init_coq_stack`, `init_coq_vm`, and `realloc_coq_stack`.

The only one leaking memory is `coq_tcode_of_code`, which is responsible for storing the bytecode before execution. I must admit that I don't understand what the point of most of this function is; is it some ultimate check that the bytecode is valid and stays so indefinitely? Anyway, there is at least one thing this function does right, it is its allocation, since it is critical that the bytecode is not stored in a block allocated by the GC!