@ppedrot 2018-03-07
> since it is critical that the bytecode is not stored in a block allocated by the GC!

Why so? Is it because it could be moved by the GC and the addresses are absolute?