@silene 2018-03-07
> Why so? Is it because it could be moved by the GC and the addresses are absolute?

Almost. The addresses stored in the bytecode are not absolute. But `coq_interp.c` was not certainly not written to account for the fact that the `pc` register and all the return addresses stored on the stack might surreptitiously become invalid. Using only relative code pointers might have a huge overhead on register-starved architectures like x86 (since it means using two registers for storing `pc` instead of a single one).