@ppedrot 2018-03-07
All right. I was planning to write a patch to allocate the bytecode through the OCaml GC using simple OCaml strings, but I can simply wrap it into an custom block instead.