@ppedrot 2018-03-07
@silene I adapted the reallocation strategy to be efficient.