@ppedrot 2018-03-07
> Almost. The addresses stored in the bytecode are not absolute. But coq_interp.c was certainly not written to account for the fact that the pc register and all the return addresses stored on the stack might surreptitiously become invalid. Using only relative code pointers might have a huge overhead on register-starved architectures like x86 (since it means using two registers for storing pc instead of a single one).

I'm slightly hijacking this thread, but as long as the GC is not called in the interpretation function, this should be safe, right? We never allocate in this function, so there cannot be any memory move nor thread rescheduling. It might be slightly dangerous if the OCaml runtime changes (e.g. because of multicore OCaml), but I guess this would already endanger the current VM...