@silene 2018-03-07
> We never allocate in this function

What gave you that idea? There are about 40 calls to `minor_collection` inside this function!