@maximedenes 2018-03-09
@ppedrot is the segfault in fiat parsers a spurious failure?