@silene 2018-03-09
Yay for automated testing. The patch is indeed wrong on several points. The removal of the following lines should have been a red flag, since if a collection happens, then the pointers to the global and atom tables will become invalid.

```
  (*action)(coq_global_data, &coq_global_data);
  (*action)(coq_atom_tbl, &coq_atom_tbl);
```