@maximedenes 2018-03-09
Ok, so postponing. And indeed, hurrah for CI!