@ppedrot 2018-03-09
Hopefully this new version is correct w.r.t. the GC kicking in. It needs to be benchmarked though.